+++ clean_dotenv/__main__.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+from clean_dotenv._main import main
+
+if __name__ == "__main__":
+    raise SystemExit(main())
