+++ setup.cfg
@@ -17,15 +17,16 @@ classifiers =
     Programming Language :: Python :: Implementation :: PyPy
 
 [options]
-py_modules = clean_dotenv
-
-install_requires =
-    python-dotenv>=1.0
+packages = find:
 python_requires = >=3.8
+
+[options.packages.find]
+exclude =
+    tests*
 
 [options.entry_points]
 console_scripts =
-    clean-dotenv = clean_dotenv:main
+    clean-dotenv = clean_dotenv._main:main
 
 [bdist_wheel]
 universal = True
